requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
